requirements.txt
kivy==2.1.0
plyer
pyjnius
pytest
